refactor export function and removed duplicate code blocks
